Overhaul feedback systems
