+++ i18n/en/code.json
@@ -628,5 +628,9 @@
   "theme.blog.author.noPosts": {
     "message": "This author has not written any posts yet.",
     "description": "The text for authors with 0 blog post"
+  },
+  "theme.feedback.helpful": {
+    "message": "Is this page helpful?",
+    "description": "The text asking if the page is helpful"
   }
 }
