+++ package.json
@@ -23,6 +23,7 @@
     "@docusaurus/preset-classic": "^3.6.3",
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
+    "@feelback/react": "^0.3.4",
     "clsx": "^1.2.1",
     "docusaurus-plugin-openapi-docs": "^4.0.1",
     "docusaurus-theme-openapi-docs": "^4.0.1",
