+++ src/theme/Footer/Rating/Rating.module.css
@@ -1,0 +1,79 @@
+/* /src/theme/Footer/Rating/Rating.module.css */
+.feedbackContainer {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.feedbackDivider {
+  width: 100%;
+  border: 0;
+  height: 1px;
+  background-color: var(--ifm-color-emphasis-200);
+  margin: 1rem 0;
+}
+
+.feedbackContent {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.questionText {
+  color: var(--ifm-color-emphasis-700);
+}
+
+.feelbackButtonsCustom :global(.feelback-buttons) {
+  display: flex;
+  gap: 0.5rem;
+}
+
+.feelbackButtonsCustom :global(.feelback-btn) {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  padding: 0.5rem 1rem;
+  border: 1px solid var(--ifm-color-primary);
+  border-radius: 0.375rem;
+  background-color: transparent;
+  color: var(--ifm-color-primary);
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.feelbackButtonsCustom :global(.feelback-btn:hover),
+.feelbackButtonsCustom :global(.feelback-btn.active) {
+  background-color: var(--ifm-color-primary);
+  color: white;
+}
+
+/* Updated SVG styling */
+.feelbackButtonsCustom :global(.feelback-btn) svg {
+  width: 1.25rem;
+  height: 1.25rem;
+  transition: transform 0.2s ease;
+}
+
+.feelbackButtonsCustom :global(.feelback-btn) svg path {
+  fill: none;
+  stroke: currentColor;
+  stroke-width: 1.5px;
+  transition: stroke-width 0.2s ease;
+}
+
+/* Modified hover selectors */
+.feelbackButtonsCustom :global(.feelback-btn:hover) svg {
+  transform: scale(1.1);
+}
+
+.feelbackButtonsCustom :global(.feelback-btn:hover) svg path {
+  stroke-width: 2px;
+}
+
+@media (max-width: 768px) {
+  .feedbackContent {
+    flex-direction: column;
+    text-align: center;
+  }
+}
