--- src/theme/Footer/custom.feedback.css
@@ -1,17 +1,0 @@
-:root {
-  --feedback-primary-color: #789806;
-  --feedback-button-dark-bg-color: var(--feedback-primary-color);
-}
-
-.feedback-wrapper {
-  visibility: hidden;
-}
-
-.feedback-wrapper.loaded {
-  visibility: visible;
-}
-
-/* Hide the feedback button initially */
-.feedback-wrapper:not(.loaded) pushfeedback-button {
-  display: none;
-}
